==> R/02_data.R <==
source("R/00_libraries.R")
source("R/01_utils.R")


# Demand data sources-----------------------------------------------------------

# population

quarterly_population_by_age_band <- quarterly_ics_populations(
  incl_agebands = TRUE
) |> 
  mutate(
    metric = paste0("Proportion of population in age band (", 
                    age_band,
                    ")"),
    frequency = "quarterly",
    quarter = case_when(
      month == 4 ~ 1L,
      month == 7 ~ 2L,
      month == 10 ~ 3L,
      month == 1 ~ 4L,
      .default = NA_integer_
    )
  ) |> 
  rename(
    numerator = "denominator"
  ) |> 
  select(!c("age_band", "month")) |> 
  mutate(
    denominator = sum(numerator),
    .by = c(
      org, year, quarter, frequency
    )
  ) |> 
  mutate(
    value = (numerator / denominator) * 100
  )
  
annual_population_by_age_band <- quarterly_to_annual_mean(
  quarterly_population_by_age_band,
  year_type = "calendar",
  multiplier = 100
)


bind_rows(
  annual_population_by_age_band,
  quarterly_population_by_age_band
) |> 
  write.csv(
    "data/population.csv",
    row.names = FALSE
    )

# risk factors
# QOF

practice_lkp <- practice_to_icb()

url <- "https://digital.nhs.uk/data-and-information/publications/statistical/quality-and-outcomes-framework-achievement-prevalence-and-exceptions-data"

zip_files <- obtain_links(url, include_link_text = TRUE) |> 
  (\(x) x[grepl("[0-9]{4}-[0-9]{2}$|\\[PAS\\]$", names(x))])() |>
  (\(x) paste0("https://digital.nhs.uk", x))() |>
  purrr::map(
    ~ obtain_links(.x, include_link_text = TRUE)
  ) |>
  purrr::map(
    .f = \(x) x[grepl("zip$", x)]
  ) |> 
  unlist() |> 
  (\(x) x[grepl("Raw", names(x))])() |> 
  (\(x) x[!grepl("1415", x)])() |> 
  # tidy up the names of the files so they are in the format "QOF yyyy-yy"
  (\(x) setNames(x, nm = gsub(":.*$", "", names(x))))()

fls <- purrr::map(
  zip_files,
  ~ download_unzip_files(
    .x, 
    directory = "data-raw/QOF",
    zip_file_pattern = "PREVALENCE",
    add_filename_prefix = TRUE
  )
) |> 
  unlist() |> 
  (\(x) setNames(
    file.rename(
      from = x,
      to = paste0(dirname(x), "/", names(x), ".csv")
    ),
    nm = paste0(dirname(x), "/", names(x), ".csv")
  ))() |> 
  enframe() |> 
  pull(name)



qof <- fls |> 
  purrr::map(
    tidy_qof
  ) |> 
  list_rbind() |> 
  left_join(
    practice_lkp,
    by = join_by(
      PRACTICE_CODE
    )
  ) |> 
  assign_org_to_missing_practices() |> 
  summarise(
    across(
      c(numerator, denominator),
      sum
    ),
    .by = c(year, org, metric, frequency)
  ) |> 
  mutate(
    value = 100 * numerator / denominator
  )

write.csv(
  qof,
  "data/qof.csv",
  row.names = FALSE
)


# risk factors from Fingertips
risk_factors <- fingertipsR::indicators() |> 
  filter(
    grepl("[Rr]isk", DomainName),
    grepl("[Ff]actor", DomainName),
    !grepl("Deprivation score \\(IMD 2019\\)", IndicatorName),
    !grepl("QOF", IndicatorName)
  ) |> 
  pull(
    IndicatorID
  ) |> 
  unique() |> 
  fingertipsR::fingertips_data(
    AreaTypeID = 221
  ) |> 
  filter(
    AreaType == "ICBs"
  ) |> 
  mutate(
    AreaCode = gsub("n", "", AreaCode)
  ) |> 
  # remove unnecessary sex/age groups from Smoking prevalence indicator
  anti_join(
    tibble(
      IndicatorID = 92443,
      Sex = c("Male", "Female", "Persons"),
      Age = c("18+ yrs", "18+ yrs", "18-64 yrs")
    ),
    by = join_by(
      IndicatorID,
      Sex,
      Age
    )
  ) |> 
  select(
    year = "Timeperiod",
    org = "AreaCode",
    org_name = "AreaName",
    metric = "IndicatorName",
    numerator = "Count",
    denominator = "Denominator",
    value = "Value",
  ) |> 
  mutate(
    frequency = case_when(
      grepl("/", year) ~ "annual financial",
      .default = "annual calendar"
    ),
    year = as.integer(
      stringr::str_extract(
        year,
        "^[0-9]{4}"
      )
    )
  ) |> 
  convert_ons_to_health_code(
    area_type = "icb",
    latest_codes_used = TRUE
  )

# if there is data already recorded in the data folder for risk factors, keep the old records that have been deleted
if (file.exists("data/risk-factors-fingertips.csv")) {
  removed_data <- read.csv("data/risk-factors-fingertips.csv") |> 
    anti_join(
      risk_factors,
      by = join_by(
        org, metric, year, frequency
      )
    )
  
  risk_factors <- risk_factors |> 
    bind_rows(
      removed_data
    )
}


risk_factors |> 
  write.csv(
    "data/risk-factors-fingertips.csv",
    row.names = FALSE
)

# deprivation
quintile_descriptions <- fingertipsR::category_types() |> 
  filter(CategoryTypeId == 6) |> 
  select(
    "Id", 
    quintile_name = "ShortName"
  )

lsoa_icb_lkp <- check_and_download(
  filepath = "data-raw/Lookups/lsoa_icb.xlsx",
  url = "https://www.arcgis.com/sharing/rest/content/items/1ac8547e9a8945478f0b5ea7ffe1a6b1/data"
) |> 
  readxl::read_excel() |> 
  select(
    "LSOA11CD",
    "ICB22CDH"
  )

url <- "https://www.ons.gov.uk/peoplepopulationandcommunity/populationandmigration/populationestimates/datasets/lowersuperoutputareamidyearpopulationestimates"
zip_files <- obtain_links(url) |> 
  (\(x) x[grepl("zip$|xlsx$", x)])() |> 
  (\(x) x[!grepl("2018on20", x)])() |> 
  (\(x) x[grepl("unformatted", x)])() |> 
  (\(x) x[!grepl("^rft-lsoa", basename(x))])() |> 
  (\(x) x[!grepl("males.zip$", x)])() |> 
  (\(x) x[!grepl("sape19|sape18", x)])() |> 
  (\(x) x[!grepl("^rft", basename(x))])()

files <- purrr::map_chr(
  paste0(
    "https://www.ons.gov.uk/",
    zip_files
  ),
  ~ check_and_download(
    url = .x,
    filepath = paste0("data-raw/Population/", basename(.x))
  )
)

pops_files <- list.files("data-raw/Population/",
                         full.names = TRUE)

pops_files <- set_names(
  pops_files,
  nm = str_extract(
    pops_files,
    "[0-9]{4}"
  )
)

lsoa_pops <- map(
  pops_files,
  open_pops_file
) |> 
  purrr::map_df(
    ~ pluck(.x, "data"),
    .id = "year"
  ) |> 
  select(
    year,
    "LSOA11CD", 
    population = "All Ages"
  ) |> 
  filter(
    grepl("^E", LSOA11CD)
  )

lsoa2122pops <- lsoa_populations_21_22() |> 
  summarise(
    population = sum(population),
    .by = c(
      year, LSOA11CD
    )
  )

lsoa_pops <- bind_rows(
  lsoa_pops,
  lsoa2122pops
)


deprivation <- check_and_download(
  filepath = "data-raw/Deprivation/deprivation.xlsx",
  url = "https://assets.publishing.service.gov.uk/media/5d8b3b51ed915d036a455aa6/File_5_-_IoD2019_Scores.xlsx"
) |> 
  readxl::read_excel(
    sheet = "IoD2019 Scores"
  ) |> 
  select(
    LSOA11CD = "LSOA code (2011)",
    imd = "Index of Multiple Deprivation (IMD) Score"
  ) |> 
  mutate(
    quintile = as.integer(6 - ntile(imd, 5))
  ) |> 
  left_join(
    quintile_descriptions,
    by = join_by(quintile == Id)
  ) |> 
  left_join(
    lsoa_icb_lkp,
    by = join_by(
      LSOA11CD
    )
  ) |> 
  left_join(
    lsoa_pops,
    by = join_by(
      LSOA11CD
    ),
    relationship = "many-to-many"
  ) |> 
  summarise(
    population = sum(population),
    .by = c(
      year,
      ICB22CDH, 
      quintile_name
    )
  ) |> 
  mutate(
    denominator = sum(population),
    .by = c(year, ICB22CDH)
  ) |> 
  rename(
    numerator = "population",
    org = "ICB22CDH"
  ) |> 
  mutate(
    value = (numerator / denominator) * 100,
    frequency = "annual calendar",
    metric = paste0(
      "Proportion of resident population in national deprivation quintile (",
      quintile_name,
      ") - IMD 2019"
    )
  ) |> 
  select(!c("quintile_name"))

write.csv(
  deprivation,
  "data/deprivation.csv",
  row.names = FALSE
)

# Capacity data sources ---------------------------------------------------

## Hospital beds
# Overnight
url <- "https://www.england.nhs.uk/statistics/statistical-work-areas/bed-availability-and-occupancy/bed-data-overnight/"

excel_urls <- obtain_links(url)

excel_urls <- excel_urls[grepl("xls$|xlsx$", excel_urls, ignore.case = TRUE)]

# download files
files <- purrr::map_chr(
  excel_urls,
  ~ check_and_download(
    url = .x,
    filepath = paste0(
      "data-raw/Bed Availability and Occupancy Data – Overnight/",
      basename(.x)
    )
  )
)

# clean filenames and remove unnecessary files
purrr::walk(
  files,
  rename_hospital_beds_xls
)

# tidy xlsx spreadsheets
# this function ignores the xls files
quarterly_overnight_beds_by_trust <- list.files(
  "data-raw/Bed Availability and Occupancy Data – Overnight/",
  full.names = TRUE
) |> 
  purrr::map_dfr(
    ~ reformat_bed_availability_data(
      filepath = .x,
      bed_type = "overnight"
    )
  )

quarterly_overnight_beds <- quarterly_overnight_beds_by_trust |> 
  apply_catchment_proportions()
  

annual_overnight_beds <- quarterly_to_annual_sum(
  quarterly_overnight_beds,
  year_type = "calendar",
  multiplier = 100
)

bind_rows(
  quarterly_overnight_beds,
  annual_overnight_beds
) |> 
  write.csv(
  "data/overnight-beds.csv",
  row.names = FALSE
)

# Day only
url <- "https://www.england.nhs.uk/statistics/statistical-work-areas/bed-availability-and-occupancy/bed-data-day-only/"
excel_urls <- obtain_links(url)
excel_urls <- excel_urls[grepl("xls$|xlsx$", excel_urls, ignore.case = TRUE)]

# download files
files <- purrr::map_chr(
  excel_urls,
  ~ check_and_download(
    url = .x,
    filepath = paste0(
      "data-raw/Bed Availability and Occupancy Data – Day/",
      basename(.x)
    )
  )
)

# clean and move file names
purrr::walk(
  files,
  rename_hospital_beds_xls
)

# tidy xlsx spreadsheets
# this function ignores the xls files
quarterly_day_beds_by_trust <- list.files(
  "data-raw/Bed Availability and Occupancy Data – Day/",
  full.names = TRUE
) |> 
  purrr::map_dfr(
    ~ reformat_bed_availability_data(
      filepath = .x,
      bed_type = "day"
    )
  )

quarterly_day_beds <- quarterly_day_beds_by_trust |> 
  apply_catchment_proportions()

annual_day_beds <- quarterly_to_annual_sum(
  quarterly_day_beds,
  year_type = "calendar",
  multiplier = 100
)


bind_rows(
  quarterly_day_beds,
  annual_day_beds
)|> 
  write.csv(
    "data/day-beds.csv",
    row.names = FALSE
  )

# total beds
total_beds <- bind_rows(
  quarterly_overnight_beds,
  annual_overnight_beds,
  quarterly_day_beds,
  annual_day_beds
) |> 
  summarise(
    across(
      c(numerator, denominator),
      sum
    ),
    .by = c(
      year, quarter, frequency, org
    )
  ) |> 
  mutate(
    value = 100 * numerator / denominator,
    metric = "Proportion of available beds that are occupied (Total)"
  )

write.csv(
  total_beds,
  "data/total-beds.csv",
  row.names = FALSE
)

# Covid hospital activity 

url <- "https://www.england.nhs.uk/statistics/statistical-work-areas/covid-19-hospital-activity/"

links <- obtain_links(url) |> 
  (\(x) x[grepl("xlsx$", x)])() |> 
  (\(x) x[!grepl("Supplementary", x)])() |> 
  (\(x) x[grepl("Covid-Publication", x)])()


files <- purrr::map_chr(
  links,
  ~ check_and_download(
    url = .x,
    filepath = paste0("data-raw/Covid activity/", basename(.x))
  )
)

# remove existing files that have been replaced online
existing_files <- list.files(
  "data-raw/Covid activity/",
  full.names = TRUE
)

setdiff(
  existing_files,
  files
) |> 
  file.remove() |> 
  invisible()

covid_numerators_by_trust <- files |> 
  map_df(
    # this function sums the numerators for each day (rather than taking a mean)
    tidy_covid_beds
  ) |> 
  # take the most recent published data where days are published more than once
  filter(
    published_date == max(published_date), 
    .by = c(
      Date
    )
  ) |> 
  summarise(
    numerator = sum(numerator),
    .by = c(
      sheet, year, quarter, org
    ),
    days = n()
  ) |> 
  arrange(sheet, org, year, quarter)

## covid bed occupancy
quarterly_total_beds_by_trust <- bind_rows(
  quarterly_day_beds_by_trust,
  quarterly_overnight_beds_by_trust
) |> 
  filter(
    grepl("Total", metric)
  ) |> 
  select(
    "year",
    "quarter",
    "org",
    "denominator"
  ) |> 
  summarise(
    denominator = sum(denominator),
    .by = c(
      year, 
      quarter,
      org
    )
  )

quarterly_covid_beds <- covid_numerators_by_trust |> 
  filter(
    sheet == "Total Beds Occupied Covid"
  ) |> 
  select(!c("sheet")) |> 
  mutate(
    numerator = numerator / days
  ) |> 
  right_join(
    quarterly_total_beds_by_trust,
    by = join_by(
      year,
      quarter,
      org
    )
) |> 
  mutate(
    numerator = replace_na(numerator, 0),
    metric = "Mean proportion of beds that contain a patient with confirmed COVID",
    frequency = "quarterly"
  ) |> 
  apply_catchment_proportions()

annual_covid_beds <- quarterly_covid_beds |> 
  quarterly_to_annual_mean(
    year_type = "calendar",
    multiplier = 100
  )

bind_rows(
  quarterly_covid_beds,
  annual_covid_beds
) |> 
  write.csv(
    "data/covid-beds.csv",
    row.names = FALSE
  )

## covid admissions
## getting total annual admissions data by trust

url <- "https://digital.nhs.uk/data-and-information/publications/statistical/hospital-admitted-patient-care-activity"

links <- obtain_links(url) |> 
  (\(x) x[grepl("[0-9]{4}-[0-9]{2}$", x)])() |> 
  (\(x) paste0("https://digital.nhs.uk", x))() |> 
  map(
    obtain_links
  ) |> 
  unlist() |> 
  (\(x) x[grepl("xlsx$", x)])() |> 
  (\(x) x[grepl("prov", x)])() |> 
  (\(x) x[!grepl("leve", x)])()

files <- links |> 
  purrr::map_chr(
    ~ check_and_download(
      filepath = paste0("data-raw/Hospital admissions/", basename(.x)),
      url = .x
    )
  )

admissions_by_trust <- files |> 
  map_df(
    tidy_admissions
  )


covid_admissions <- covid_numerators_by_trust |> 
  filter(
    sheet == "Admissions Total"
  ) |> 
  mutate(
    org = case_when(
      grepl("^R", org) & (nchar(org) == 5) ~ substr(org, 1, 3),
      .default = org
    )
  ) |> 
  summarise(
    numerator = sum(numerator),
    .by = c(
      year, org
    )
  )

# Org codes in covid_admissions and not in admissions_by_trust are community
# hospitals. Org codes in admissions_by_trust and not in covid_admissions are
# "treatment centres" (R0176), which often includes eye care
admissions_with_covid <- covid_admissions |> 
  inner_join(
    admissions_by_trust,
    by = join_by(org, year)
  ) |> 
  mutate(
    frequency = "annual financial",
    metric = "Mean proportion of admissions with COVID diagnosis in last 24 hours"
  ) |> 
  apply_catchment_proportions()

backcast_denominator <- admissions_with_covid |> 
  filter(year == min(year)) |> 
  select(
    !c("year", "numerator", "value")
  ) |> 
  cross_join(
    tibble(
      year = seq(
        from = min(admissions_by_trust$year),
        to = min(admissions_with_covid$year) - 1,
        by = 1
      ),
      numerator = 0,
      value = 0
    )
  )

admissions_with_covid |> 
  bind_rows(
    backcast_denominator
  ) |> 
  mutate(
    value = value * 100
  ) |> 
  write.csv(
    "data/covid-admissions.csv",
    row.names = FALSE
  )
  
## covid absences
workforce <- tidy_workforce_ftes(
  type = "HC" # headcount
) |> 
  rename(
    denominator = "numerator"
  ) |> 
  filter(
    grepl("^Total", metric)
  )

covid_absences <- covid_numerators_by_trust |> 
  filter(
    sheet == "Covid Absences"
  ) |> 
  summarise(
    numerator = sum(numerator / days),
    .by = c(
      year, org
    )
  ) |> 
  inner_join(
    workforce,
    by = join_by(
      org, year
    )
  ) |> 
  select(!c("month", "metric")) |> 
  apply_catchment_proportions() |> 
  mutate(
    frequency = "annual financial",
    metric = "Mean proportion of workforce absent in a provider setting due to COVID per available headcount"
  )

backcast_denominator <- covid_absences |> 
  filter(year == min(year)) |> 
  select(
    !c("year", "numerator", "value")
  ) |> 
  cross_join(
    tibble(
      year = seq(
        from = min(workforce$year),
        to = min(covid_absences$year) - 1,
        by = 1
      ),
      numerator = 0,
      value = 0
    )
  )

covid_absences |> 
  bind_rows(
    backcast_denominator
  ) |> 
  mutate(
    value = value * 100
  ) |> 
  write.csv(
    "data/covid-absences.csv",
    row.names = FALSE
  )

# Total beds per 60+ population

population_60_plus <- read.csv(
  "data/population.csv"
) |> 
  filter(
    grepl(paste(seq(60, 90, by = 10), collapse = "|"),
          metric)
  ) |> 
  summarise(
    denominator = sum(numerator),
    .by = c(
      year, org
    )
  )

quarterly_beds_per_60plus <- c(
  "data/overnight-beds.csv",
  "data/day-beds.csv",
  "data/total-beds.csv"
) |> 
  purrr::map_dfr(
    read.csv
  ) |> 
  mutate(
    metric = case_when(
      metric == "Proportion of available beds that are occupied (Total)" ~ "Total beds per 1,000 60+ yrs (Total)",
      grepl("General", metric) ~ gsub(
        "Proportion of available beds that are occupied",
        "Total beds per 1,000 60+ yrs",
        metric
      ),
      .default = NA_character_
    )
  ) |> 
  filter(
    frequency == "quarterly",
    !is.na(metric)
  ) |> 
  select(
    "year", "quarter",
    "org", "frequency",
    numerator = "denominator",
    metric
  ) |> 
  inner_join(
    population_60_plus,
    by = join_by(
      year, org
    ),
    relationship = "many-to-one"
  ) |> 
  mutate(
    value = 1e3 * numerator / denominator
  )

annual_beds_per_60plus <- quarterly_beds_per_60plus |> 
  quarterly_to_annual_mean(
    year_type = "calendar",
    multiplier = 1e3
  )

bind_rows(
  quarterly_beds_per_60plus,
  annual_beds_per_60plus
) |> 
  write.csv(
    "data/beds-per-1000-60plus.csv",
    row.names = FALSE
  )


# NHS workforce per population
# collecting the numerator
annual_workforce_fte <- tidy_workforce_ftes() |> 
  apply_catchment_proportions()

# collecting the denominator (populations by health areas)
quarterly_health_pop_denominators <- quarterly_ics_populations()

workforce_metrics <- annual_workforce_fte |> 
  complete(
    year, month, org, metric
  ) |> 
  inner_join(
    quarterly_health_pop_denominators,
    by = join_by(
      org,
      year, 
      month
    )
  ) |> 
  mutate(
    numerator = replace_na(numerator, 0),
    metric = paste0(
      "Workforce FTEs per 100,000 population (",
      metric,
      ")"
    ),
    value = (numerator / denominator) * 1e5,
    frequency = "annual calendar"
  ) |> 
  select(!c("month"))

write.csv(
  workforce_metrics,
  "data/nhs_workforce_metrics.csv",
  row.names = FALSE
)

# a&e workforce
url <- "https://digital.nhs.uk/data-and-information/publications/statistical/hospital-accident--emergency-activity/"

links <- obtain_links(
  url,
  include_link_text = TRUE
) |> 
  (\(x) x[grepl("[0-9]{4}-[0-9]{2}", names(x))])() |> 
  (\(x) paste0("https://digital.nhs.uk", x))() |> 
  map(
    ~ obtain_links(
      .x,
      include_link_text = TRUE
    )
  ) |> 
  unlist() |> 
  (\(x) x[grepl("workforce|provider", names(x), ignore.case = TRUE)])() |>
  enframe(
    name = "link_text",
    value = "hyperlink"
  ) |> 
  mutate(
    yr = str_extract(
      string = link_text,
      pattern = "[0-9]{4}-[0-9]{2}"
    ),
    filetype = tools::file_ext(hyperlink)
  ) |> 
  filter(
    filetype %in% c("xlsx", "xls", "csv"),
    !grepl("ECDS|DQ", link_text),
    as.numeric(substr(yr, 1, 4)) > 2016
  ) |> 
  mutate(
    cnt = n(),
    .by = yr
  ) |> 
  mutate(
    type = case_when(
      grepl("Workforce", link_text) ~ "Workforce",
      .default = "Provider"
    )
  ) |> 
  filter(
    cnt == 1 |
      type == "Workforce"
  ) |> 
  pull(hyperlink)


files <- purrr::map_chr(
  links,
  ~ check_and_download(
    url = .x,
    filepath = paste0("data-raw/A and E workforce/", basename(.x))
  )
) 

a_and_e_workforce_quarterly <- files |> 
  map_df(
    tidy_a_and_e_workforce, type = "fte"
  ) |> 
  apply_catchment_proportions()

# collecting the denominator (populations by health areas)
quarterly_health_pop_denominators <- quarterly_ics_populations() |> 
  mutate(
    quarter = case_when(
      month == 4 ~ 1L,
      month == 7 ~ 2L,
      month == 10 ~ 3L,
      month == 1 ~ 4L,
      .default = NA_integer_
    ),
    year = case_when(
      quarter == 4L ~ year - 1,
      .default = year
    )
  ) |> 
  select(!c("month"))

a_and_e_workforce_quarterly_per_population <- a_and_e_workforce_quarterly |> 
  left_join(
    quarterly_health_pop_denominators,
    by = join_by(
      year, org, quarter
    )
  ) |> 
  mutate(
    value = numerator / (denominator / 1e5),
    metric = paste0(metric, " (per 100,000 population)")
  )

a_and_e_workforce_annual_per_population <- quarterly_to_annual_mean(
  a_and_e_workforce_quarterly_per_population,
  year_type = "calendar",
  multiplier = 1e5
)

bind_rows(
  a_and_e_workforce_quarterly_per_population,
  a_and_e_workforce_annual_per_population
) |> 
  write.csv(
    "data/a-and-e-workforce.csv",
    row.names = FALSE
  )

# Primary care workforce
url <- "https://digital.nhs.uk/data-and-information/publications/statistical/general-and-personal-medical-services"

links <- obtain_links(
  url,
  include_link_text = TRUE) |> 
  (\(x) x[grepl("General Practice Workforce", names(x))])() |> 
  (\(x) x[grepl("[0-9]{4}$", names(x))])() |> 
  (\(x) paste0("https://digital.nhs.uk", x))() |> 
  map(
    ~ obtain_links(
      .x,
      include_link_text = TRUE
    )
  ) |> 
  unlist() |> 
  # (\(x) x[grepl("xls$|xlsx$|zip$", x)])() |> 
  (\(x) x[grepl("Individual", names(x))])() |> 
  (\(x) x[!grepl("Practice", names(x))])() |> 
  (\(x) x[grepl("March|June|September|December", names(x))])() |> 
  (\(x) x[!grepl("March 2022|June 2022|December 2021", names(x))])()

gp_workforce_directory <- "data-raw/GP workforce"

files <- map(
  links,
  ~ download_unzip_files(
    zip_url = .x,
    directory = gp_workforce_directory,
    zip_file_pattern = ".*Individual.*csv$"
  )
)

quarterly_gp_workforce <- tibble(
  fname = list.files(gp_workforce_directory)
) |> 
  mutate(
    mnth = str_extract(fname, paste(month.name, collapse = "|")),
    mnth_num = match(mnth, month.name),
    yr = str_extract(fname, "[0-9]{4}")
  ) |> 
  filter(
    mnth_num %in% c(3, 6, 9, 12),
    grepl("^[0-9]", fname)) |> 
  filter(
    n() == 4,
    .by = yr
  ) |> 
  pull(fname) |> 
  (\(x) paste0(gp_workforce_directory, "/", x))() |> 
  map(
    read.csv
  ) |> 
  map(
    ~ mutate(
      .x,
      STAFF_GROUP = case_when(
        STAFF_GROUP == "GP" ~ STAFF_GROUP,
        .default = str_to_title(STAFF_GROUP)
      )
    )
  ) |> 
  map(
    ~ summarise(
      .x, 
      numerator = sum(FTE),
      .by = any_of(
        c(
          "YEAR",
          "Month",
          "ICB_CODE",
          "STP_CODE",
          "STAFF_GROUP"#,
          #"STAFF_ROLE"
        )
      )
    )
  ) |> 
  map(
    ~ rename(
      .x,
      org = any_of(
        c("STP_CODE", "ICB_CODE")
      ),
      year = "YEAR",
      month = "Month"
    )
  ) |> 
  bind_rows()

health_populations <- quarterly_ics_populations() |> 
  mutate(
    month = month + 2
  )


quarterly_gp_workforce_per_population <- quarterly_gp_workforce |> 
  inner_join(
    health_populations,
    by = join_by(
      org, year, month
    ),
    relationship = "many-to-one"
  ) |> 
  mutate(
    value = numerator / (denominator / 1e5),
    metric = paste0(
      "Primary care workforce (FTEs) per 100,000 population (",
      STAFF_GROUP,
      # " - ", 
      # STAFF_ROLE,
      ")"),
    frequency = "quarterly",
    quarter = str_replace_all(
      month,
      c(
        `3` = "4",
        `6` = "1",
        `9` = "2",
        `12` = "3"
      )
    ),
    quarter = as.integer(quarter)
  ) |> 
  select(!c(
    "STAFF_GROUP", 
    "month"
    # "STAFF_ROLE"
  ))

annual_gp_workforce_per_population <- quarterly_to_annual_mean(
  quarterly_gp_workforce_per_population,
  year_type = "calendar",
  multiplier = 1e5
)

bind_rows(
  quarterly_gp_workforce_per_population,
  annual_gp_workforce_per_population
) |> 
  write.csv(
    "data/primary-care-workforce-per-population.csv",
    row.names = FALSE
  )

# sickness absence
# collecting numerators
# April 2009 to March 2022 file
url <- "https://files.digital.nhs.uk/A6/67E0C1/ESR_ABSENCE_CSV_NHSE.csv"

monthly_sickness_absence <- read.csv(url) |> 
  mutate(
    year = as.integer(
      substr(
        Date, 1, 4
        )
      ),
    month = match(
      tolower(
        gsub(".*-", "", Date)
      ),
      tolower(month.abb)
    ),
    Org.Type = tolower(Org.Type),
    frequency = "monthly"
  ) |> 
  select(
    "year",
    "month",
    org = "Org.Code",
    org_name = "Org.Name",
    `FTE days sick` = "FTE.Days.Sick",
    `FTE days available` = "FTE.Days.Available",
    Org.Type,
    "frequency"
  ) |> 
  pivot_longer(
    cols = starts_with("FTE"),
    names_to = "metric",
    values_to = "numerator"
  ) |> 
  mutate(
    metric = case_when(
      grepl("sick", metric) ~ paste0(metric, " per 10,000 population (",
                                     Org.Type, ")"),
      .default = paste0(metric, " per 100 population (",
                        Org.Type, ")")
    )
  ) |> 
  select(
    !c("Org.Type")
  )

# monthly files April 2022 onwards
url <- "https://digital.nhs.uk/data-and-information/publications/statistical/nhs-sickness-absence-rates"

links <- obtain_links(url) |> 
  (\(x) x[grepl(tolower(paste(month.name, collapse = "|")), x)])() |> 
  (\(x) x[!grepl(tolower(paste(2014:2021, collapse = "|")), x)])() |> 
  (\(x) x[!grepl(tolower(paste(c("january", "february"), 2022, 
                               sep = "-", 
                               collapse = "|")), x)])()

csv_links <- purrr::map(
  .x = paste0(
    "https://digital.nhs.uk/",
    links
  ),
  .f = obtain_links
) |> 
  unlist() |> 
  (\(x) x[grepl("csv$", x)])() |> 
  (\(x) x[!grepl("benchmarking|reason|COVID|ESR_ABSENCE_CSV_NHSE", x, ignore.case = TRUE)])()

monthly_sickness_monthly_files <- purrr::map_dfr(
  csv_links,
  read.csv
) |> 
  mutate(
    year = as.integer(
      substr(
        DATE, nchar(DATE) - 3, nchar(DATE)
      )
    ),
    month = as.integer(
      substr(
        DATE, nchar(DATE) - 6, nchar(DATE) - 5
      )
    ),
    ORG_TYPE = tolower(ORG_TYPE),
    frequency = "monthly"
  ) |> 
  select(
    "year",
    "month",
    org = "ORG_CODE",
    org_name = "ORG_NAME",
    "ORG_TYPE",
    `FTE days sick` = "FTE_DAYS_LOST",
    `FTE days available` = "FTE_DAYS_AVAILABLE",
    "frequency"
  ) |> 
  pivot_longer(
    cols = starts_with("FTE"),
    names_to = "metric",
    values_to = "numerator"
  ) |> 
  mutate(
    metric = case_when(
      grepl("sick", metric) ~ paste0(metric, " per 10,000 population (",
                                     ORG_TYPE, ")"),
      .default = paste0(metric, " per 100 population (",
                        ORG_TYPE, ")")
    )
  ) |> 
  select(
    !c("ORG_TYPE")
  )

monthly_sickness_absence <- bind_rows(
    monthly_sickness_absence, 
    monthly_sickness_monthly_files
  )

trust_ics_lkp <- trust_to_ics_proportions(
  final_year = max(monthly_sickness_absence$year)
) |> 
  rename(
    health_org_code = "TrustCode",
    icb_code = "org"
  )

orgs <- monthly_sickness_absence |> 
  distinct(org) |> 
  mutate(
    role = map_chr(
      .x = org,
      .f = health_org_role
    )
  ) |> 
  filter(
    role %in% c(
      # "RO98" CCG 200
      "RO107", # care trust 11
      # "RO108" # care trust sites 0
      # "RO116" Executive agency programme 3
      "RO157", # Independent providers 4
      # "RO162" Executive agency programme 3
      "RO172", # Independent Sector Healthcare providers 32
      # "RO189" Special Health Authorities 13
      "RO197" # NHS trust 259
      # "RO198", # NHS trust sites 0
      # "RO213" Commissioning Support Units 32
      # "RO261" Executive Agency Programme 42
    )
  ) |> 
  pull(org) |> 
  setdiff(
    unique(trust_ics_lkp$health_org_code)
  )

org_lkp <- nearest_health_orgs(
  missing_orgs = orgs,
  known_orgs = unique(trust_ics_lkp$health_org_code),
  n = 2
) |> 
  mutate(
    known_org_proportion = 1 - (distance / sum(distance)),
    .by = missing_org
  ) |> 
  left_join(
    trust_ics_lkp,
    by = join_by(
      known_org == health_org_code
    ),
    relationship = "many-to-many"
  ) |> 
  mutate(
    proportion = proportion * known_org_proportion
  ) |> 
  select(
    "year",
    health_org_code = "missing_org",
    "icb_code",
    "proportion"
  ) |> 
  bind_rows(
    trust_ics_lkp
  )

# denominators
quarterly_health_pop_denominators <- quarterly_ics_populations() |> 
  rename(
    quarter = "month"
  ) |> 
  mutate(
    quarter = case_when(
      quarter == 1 ~ 4L,
      quarter == 4 ~ 1L,
      quarter == 7 ~ 2L,
      quarter == 10 ~ 3L,
      .default = NA_real_
    )
  )

quarterly_sickness_absence <- monthly_sickness_absence |> 
  inner_join(
    org_lkp,
    by = join_by(
      year,
      org == health_org_code
    ),
    relationship = "many-to-many"
  ) |> 
  mutate(
    quarter = case_when(
      month %in% 1:3 ~ 4L,
      month %in% 4:6 ~ 1L,
      month %in% 7:9 ~ 2L,
      month %in% 10:12 ~ 3L,
      .default = NA_real_
    ),
    year = case_when(
      quarter == 4 ~ year - 1,
      .default = year
    )
  ) |> 
  summarise(
    across(
      c(numerator),
      ~ sum(.x * proportion, na.rm = TRUE)
    ),
    .by = c(
      year, quarter, metric, icb_code
    )
  ) |> 
  inner_join(
    quarterly_health_pop_denominators,
    by = join_by(
      icb_code == org,
      year,
      quarter
    )
  ) |> 
  mutate(
    value = case_when(
      grepl("sick", metric) ~ (numerator / denominator) * 1e4,
      .default = (numerator / denominator) * 1e2
    ),
    frequency = "quarterly"
  ) |> 
  rename(
    org = "icb_code"
  )

annual_sickness_absence <- quarterly_to_annual_mean(
  quarterly_sickness_absence,
  year_type = "calendar"
) |> 
  mutate(
    value = case_when(
      grepl("sick", metric) ~ value * 1e4,
      .default = value * 1e2
    )
  )

bind_rows(
  quarterly_sickness_absence,
  annual_sickness_absence
) |> 
  write.csv(
    "data/sickness-absence.csv",
    row.names = FALSE
    )


# NHS payments to General Practices
# https://digital.nhs.uk/data-and-information/publications/statistical/nhs-payments-to-general-practice


# Mental health spend data

url <- "https://www.england.nhs.uk/publication/nhs-mental-health-dashboard/"

excel_link <- obtain_links(url) |> 
  (\(x) x[grepl("xlsm$|xlsb$", x)])() |> 
  as.character() |> 
  head(1)

list.files("data-raw/Mental Health Spend/",
           full.names = TRUE) |> 
  file.remove() |> 
  invisible()

xl_file <- check_and_download(
  filepath = paste0("data-raw/Mental Health Spend/", basename(excel_link)),
  url = excel_link
)

annual_mh_spend <- tidy_mh_file(xl_file)

# annual health population denominators

annual_health_pop_denominators <- quarterly_ics_populations() |> 
  mutate(
    year = case_when(
      month == 1 ~ year - 1,
      .default = year
    )
  ) |> 
  summarise(
    denominator = mean(denominator),
    .by = c(
      org, year
    )
  )

annual_mh_spend_metrics <- annual_mh_spend |> 
  inner_join(
    annual_health_pop_denominators,
    by = join_by(
      org, year
    ),
    relationship = "many-to-one"
  ) |> 
  mutate(
    denominator = case_when(
      grepl("proportion", metric) ~ 1,
      .default = denominator
    ),
    value = case_when(
      grepl("allocation", metric) ~ (numerator / denominator) * 100,
      .default = numerator / denominator
    )
  )

write.csv(
  annual_mh_spend_metrics,
  "data/mh-spend.csv",
  row.names = FALSE
)

# Performance -------------------------------------------------------------

# For interpretation, the performance metrics are calculated so the higher they
# are the worse the metric


# GP wait times these are help in zip files containing monthly of data (each zip
# file contains multiple csvs, one for each month). The zip files are produced
# monthly, so there is overlap between the csv files in multiple zip files, but
# the data changes slightly (as information updates?). This function will always
# use the most recent file for each month for its information
url <- "https://digital.nhs.uk/data-and-information/publications/statistical/appointments-in-general-practice/"

url_links <- obtain_links(url) |> 
  (function(url) url[grepl("[a-z].*-[0-9]{4}$", url)])() |> 
  (function(url) paste0("https://digital.nhs.uk", url))() |> 
  purrr::map(
    obtain_links
  ) |> 
  lapply(
    function(x) x[grepl("zip$", x)]
  ) |> 
  lapply(
    function(x) if (length(x) > 1) {
      x[grepl("Daily", x)]
    } else {
      x
    }
  ) |> 
  unlist() |> 
  (\(x) x[grepl("[a-z]{3}_{0,1}[0-9]{2}", x, ignore.case = TRUE)])() |> 
  (\(x) x[grepl(paste(month.abb, collapse = "|"), x, ignore.case = TRUE)])() |> 
  (\(x) x[seq(1, length(x), 4)])()

list.files("data-raw/GP wait times/",
           full.names = TRUE) |> 
  file.remove() |> 
  invisible()

files <- purrr::map(
  url_links,
  download_unzip_gp_wait_times
)

monthly_gp_wait_times <- setNames(
  list.files(
    "data-raw/GP wait times/"
  ),
  nm = gsub(".csv", "", list.files(
    "data-raw/GP wait times/"
  ))) |> 
  (\(x) x[!grepl("Daily", x)])() %>% 
  purrr::map_dfr(
    ~ read.csv(
      paste0(
        "data-raw/GP wait times/",
        .x
      ),
      colClasses = c(
        "character",
        "character",
        "integer"
      )
    ),
    .id = "date"
  ) |> 
  rename(
    org = SUB_ICB_CODE
  ) |> 
  update_to_latest_ics_codes() |> 
  mutate(
    year = as.integer(
      paste0(
        "20",
        substr(
          date, 5, 6
        )
      )
    ),
    month = match(
      substr(
        date, 1, 3
      ),
      month.abb
    ),
    TIME_BETWEEN_BOOK_AND_APPT = case_when(
      TIME_BETWEEN_BOOK_AND_APPT %in% c("Same Day", "1 Day", "2 to 7 Days") ~ "0 to 1 weeks",
      TIME_BETWEEN_BOOK_AND_APPT %in% c("8  to 14 Days") ~ "1 to 2 weeks",
      TIME_BETWEEN_BOOK_AND_APPT %in% c("15  to 21 Days") ~ "2 to 3 weeks",
      TIME_BETWEEN_BOOK_AND_APPT %in% c("22  to 28 Days") ~ "3 to 4 weeks",
      TIME_BETWEEN_BOOK_AND_APPT %in% c("More than 28 Days") ~ "over 4 weeks",
      .default = "Unknown/Data Issue"
    ),
    metric = paste0(
      "Proportion of attended GP appointments",
      " (",
      TIME_BETWEEN_BOOK_AND_APPT,
      " wait time)"
    )
  ) |> 
  filter(
    TIME_BETWEEN_BOOK_AND_APPT != "Unknown/Data Issue"
  ) |>
  summarise(
    numerator = sum(COUNT_OF_APPOINTMENTS),
    .by = c(
      date,
      year,
      month,
      org,
      metric,
      TIME_BETWEEN_BOOK_AND_APPT
    )
  ) |> 
  mutate(
    denominator = sum(numerator),
    .by = c(
      date,
      year, 
      month,
      org
    )
  ) |> 
  mutate(
    value = (numerator / denominator) * 100,
    frequency = "monthly"
  ) |> 
  select(
    "year",
    "month",
    "org",
    "metric",
    "numerator",
    "denominator",
    "value",
    "frequency"
  )

quarterly_gp_wait_times <- monthly_to_quarterly_sum(
  monthly_gp_wait_times,
  multiplier = 100
)

annual_gp_wait_times <- monthly_to_annual_sum(
  monthly_gp_wait_times,
  year_type = "calendar",
  multiplier = 100
)


bind_rows(
  monthly_gp_wait_times,
  quarterly_gp_wait_times,
  annual_gp_wait_times
) |> 
  write.csv(
    "data/gp-wait-times.csv",
    row.names = FALSE
    )  

# ambulance response times
url <- "https://www.england.nhs.uk/statistics/wp-content/uploads/sites/2/2023/09/20230825-AmbSYS-AmbCO-indicator-list.xlsx"

file <- check_and_download(
  url = url,
  filepath = paste0("data-raw/Ambulance response times/", basename(url))
)

keep_indicators <- paste0("A", c(8:12, 24, 27, 30, 33, 36))

ambsys_lookup <- read_excel(
  file,
  sheet = "AmbSYS",
  range = "R5C1:R136C3",
  col_names = c(
    "id",
    "description",
    "metric"
  )
) |> 
  filter(
    id %in% keep_indicators
  )

csv_link <- obtain_links("https://www.england.nhs.uk/statistics/statistical-work-areas/ambulance-quality-indicators/") |> 
  (function(x) x[grepl("csv$", x)])() |> 
  (function(x) x[grepl("^AmbSYS", basename(x))])()

file <- check_and_download(
  url = csv_link,
  filepath = paste0("data-raw/Ambulance response times/", basename(csv_link))
)

monthly_ambsys <- read.csv(
  file,
  na.strings = "."
) |> 
  tidyr::pivot_longer(
    cols = starts_with("A"),
    names_to = "id",
    values_to = "value", 
    values_transform = function(x) gsub(",", "", x)
  ) |> 
  inner_join(
    ambsys_lookup,
    by = join_by(id)
  ) |> 
  mutate(
    description = gsub(" response", "", description)
  ) |> 
  select(!c("id")) |> 
  tidyr::pivot_wider(
    names_from = metric,
    values_from = value
  ) |> 
  rename(
    numerator = "total time",
    denominator = "incident count"
  ) |> 
  mutate(
    across(
      c(numerator, denominator),
      as.numeric
    ),
    numerator = numerator / 60, # convert from seconds to minutes
    metric = paste0(
      "Mean ambulance response time - minutes (",
      description,
      ")"
    ),
    value = numerator / denominator,
    frequency = "monthly"
  ) |> 
  filter(
    !is.na(numerator),
    !is.na(denominator)
  ) |> 
  select(
    year = "Year",
    month = "Month",
    org = "Org.Code",
    org_name = "Org.Name",
    "metric",
    "value",
    "numerator",
    "denominator",
    "frequency"
  )

quarterly_ambsys <- monthly_to_quarterly_sum(
  monthly_ambsys
)

annual_ambsys <- monthly_to_annual_sum(
  monthly_ambsys,
  year_type = "calendar"
)

bind_rows(
  monthly_ambsys,
  quarterly_ambsys,
  annual_ambsys
) |> 
  write.csv(
    "data/ambulance-response-times.csv",
    row.names = FALSE
)

# No criteria to reside
url <- "https://www.england.nhs.uk/statistics/statistical-work-areas/discharge-delays-acute-data/"
excel_links <- obtain_links(url) |> 
  (\(x) x[grepl("xlsx$", x)])() |> 
  (\(x) x[!grepl("[[:alpha:]].*[0-9]{4}-[[:alpha:]].*[0-9]{4}", x)])() |> 
  (\(x) x[grepl("sitrep", x)])()

files <- purrr::map_chr(
  excel_links,
  ~ check_and_download(
    filepath = paste0("data-raw/No criteria to reside/", basename(.x)),
    url = .x
  )
)

tidy_nctr <- purrr::map(
  files,
  tidy_nctr_file
)

monthly_nctr <- purrr::map_dfr(
  tidy_nctr,
  aggregate_nctr_to_month
) |> 
  apply_catchment_proportions(
    proportion = TRUE
  )

quarterly_nctr <- monthly_to_quarterly_sum(
  monthly_nctr,
  multiplier = 100
)

annual_nctr <- monthly_to_annual_sum(
  monthly_nctr,
  year_type = "calendar",
  multiplier = 100
)

bind_rows(
  monthly_nctr,
  quarterly_nctr,
  annual_nctr
) |> 
  write.csv(
    "data/no-criteria-to-reside.csv",
    row.names = FALSE
  )

# 62 day cancer waiting times
url <- "https://www.england.nhs.uk/statistics/statistical-work-areas/cancer-waiting-times/"
excel_links <- obtain_links(url) |> 
  (\(x) x[grepl("xlsx$", x)])() |> 
  (\(x) x[grepl("Commissioner", x)])() |> 
  (\(x) x[grepl("Revision", x)])() |> 
  (\(x) x[!grepl("Jul-2022-Sep-2023", x)])()

list.files("data-raw/Cancer wait times/", full.names = TRUE) |> 
  file.remove() |> 
  invisible()

files <- purrr::map_chr(
  excel_links,
  ~ check_and_download(
    filepath = paste0("data-raw/Cancer wait times/", basename(.x)),
    url = .x
  )
)

monthly_cancer_wait_times <- purrr::map_dfr(
  files,
  tidy_cancer_wait_times
)

orgs <- monthly_cancer_wait_times |> 
  pull(org) |> 
  unique() |> 
  (\(x) x[!grepl("UNKNOWN", x)])()
  
org_lkp <- orgs |> 
  attach_icb_to_org()

monthly_cancer_wait_times <- monthly_cancer_wait_times |> 
  filter(
    # remove welsh organisations and Unknown
    !(org %in% c(
      paste0("7A", 1:7),
      "6A8",
      "UNKNOWN"
    ))
  ) |> 
  left_join(
    org_lkp,
    by = join_by(org == health_org_code)
  ) |> 
  summarise(
    across(
      c(numerator, denominator),
      ~ sum(.x, na.rm = TRUE)
    ),
    .by = c(
      year, month, icb_code, metric, frequency
    )
  ) |> 
  rename(
    org = icb_code
  ) |> 
  mutate(
    value = (numerator / denominator) * 100
  )

quarterly_cancer_wait_times <- monthly_to_quarterly_sum(
  monthly_cancer_wait_times,
  multiplier = 100
)

annual_cancer_wait_times <- monthly_to_annual_sum(
  monthly_cancer_wait_times,
  year_type = "calendar",
  multiplier = 100
)

bind_rows(
  monthly_cancer_wait_times,
  quarterly_cancer_wait_times,
  annual_cancer_wait_times
) |> 
  write.csv(
    "data/62-day-cancer-wait-times.csv",
    row.names = FALSE
  )

# A&E 4 hour waits

url <- "https://www.england.nhs.uk/statistics/statistical-work-areas/ae-waiting-times-and-activity/"

monthly_data_links <- obtain_links(url) |> 
  (function(x) x[grepl("ae-attendances", basename(x))])() |> 
  unique()

excel_file_links <- purrr::map(
  monthly_data_links,
  obtain_links
  ) |> 
  unlist() |> 
  (function(x) x[grepl("xls$|xlsx$", x)])() |> 
  (function(x) x[grepl(paste(month.name, collapse = "|"),
                       basename(x))])() |> 
  # (function(x) x[grepl("by-provider", x)])()
  (function(x) x[!grepl("Growth", x)])() |> 
  (function(x) x[!grepl("Quarter", x)])() |> 
  (function(x) x[!grepl("Supplementary", x)])() |> 
  (\(x) x[!grepl("Attribution", x)])()

files <- purrr::map_chr(
  excel_file_links,
  ~ check_and_download(
    filepath = paste0("data-raw/A and E wait times/", basename(.x)),
    url = .x
  )
)

monthly_a_and_e <- purrr::map_dfr(
  files, 
  tidy_a_and_e
) |> 
  apply_catchment_proportions()

quarterly_a_and_e <- monthly_to_quarterly_sum(
  monthly_a_and_e,
  multiplier = 100
)

annual_a_and_e <- monthly_to_annual_sum(
  monthly_a_and_e,
  multiplier = 100,
  year_type = "calendar"
)  

bind_rows(
  monthly_a_and_e,
  quarterly_a_and_e,
  annual_a_and_e
) |> 
  write.csv(
    "data/a-and-e-4-hours.csv",
    row.names = FALSE
  )

# referral to treatment
url <- "https://www.england.nhs.uk/statistics/statistical-work-areas/rtt-waiting-times/"
annual_urls <- obtain_links(url) |> 
  (\(x) x[grepl("^https://www.england.nhs.uk/statistics/statistical-work-areas/rtt-waiting-times/", x)])() |> 
  (\(x) x[grep("[0-9]{4}-[0-9]{2}", x)])() |> 
  unique() |> 
  (\(x) x[!grepl("2011-12|2012-13", x)])()

xl_files <- purrr::map(
  annual_urls,
  obtain_links
) |> 
  unlist() |> 
  (function(x) x[grepl("xls$|xlsx$", x)])() |> 
  (function(x) x[grepl("Commissioner", x)])() |> 
  (function(x) x[grepl("Incomplete", x)])()

files <- purrr::map_chr(
  xl_files,
  ~ check_and_download(
    filepath = paste0("data-raw/Referral to Treatment/", basename(.x)),
    url = .x
  )
) |> 
  purrr::map_chr(
    rename_rtt_files
  )

# obtain new list of files in the folder
files <- list.files(
  "data-raw/Referral to Treatment",
  full.names = TRUE
)


# tidy the data in the files
monthly_rtt_18 <- files |> 
  purrr::map_dfr(
    ~ tidy_rtt(
      .x,
      num_weeks = 18
    )
  )

monthly_rtt_52 <- files |> 
  purrr::map_dfr(
    ~ tidy_rtt(
      .x,
      num_weeks = 52
    )
  )
monthly_rtt_65 <- files |> 
  purrr::map_dfr(
    ~ tidy_rtt(
      .x,
      num_weeks = 65
    )
  )

monthly_rtt <- bind_rows(
  monthly_rtt_18,
  monthly_rtt_52,
  monthly_rtt_65
)

org_icb_lkp <- monthly_rtt |> 
  distinct(
    org
  ) |> 
  filter(
    org != "-"
  ) |> 
  pull() |> 
  attach_icb_to_org()

monthly_rtt <- monthly_rtt |> 
  inner_join(
    org_icb_lkp,
    by = join_by(
      org == health_org_code
    )
  ) |> 
  summarise(
    across(
      c(numerator, denominator),
      ~ sum(.x, na.rm = TRUE)
    ),
    .by = c(
      year, month, icb_code, metric, frequency
    )
  ) |> 
  rename(
    org = "icb_code"
  ) |> 
  mutate(
    value = (numerator / denominator) * 100
  )

quarterly_rtt <- monthly_to_quarterly_sum(
  monthly_rtt,
  multiplier = 100
)

annual_rtt <- monthly_to_annual_sum(
  monthly_rtt,
  year_type = "calendar",
  multiplier = 100
)

bind_rows(
  monthly_rtt,
  quarterly_rtt,
  annual_rtt
) |> 
  write.csv(
    "data/referral-to-treatment.csv",
    row.names = FALSE
  )
